Fix some regressions

+++ src/wirecloud/platform/workspace/models.py
@@ -65,7 +65,7 @@ class Workspace(models.Model):
             update_workspace_preferences(self, {'public': {'value': self.public}})
             self.__original_public = self.public
 
-        self.last_modified = time.time() * 1000
+        self.last_modified = long(time.time() * 1000)
 
         super(Workspace, self).save(*args, **kwargs)
